requirements.txt
streamlit
feedparser
pandas
folium
streamlit-folium
geopy
python-telegram-bot==13.15
